test_requirements.txt
djangocms-helper
pysqlite
pytz
coverage
